src/StateBasedTeamsAgent/requirements.txt
# Azure AI Agent Server for hosted agents
azure-ai-agentserver-agentframework
agent-framework

# Azure AI integration (versions must be compatible with agentserver)
azure-ai-projects>=1.0.0b11
azure-ai-inference
azure-identity

# Document Intelligence (PDF extraction)
azure-ai-documentintelligence>=1.0.0

# AI Language (PII removal)
azure-ai-textanalytics>=5.3.0

# Additional utilities
python-dotenv
pydantic
pydantic-settings

# Original dependencies  
openai

# Agent SDK dependencies
aiohttp
microsoft-agents-hosting-aiohttp
microsoft-agents-hosting-core
microsoft-agents-authentication-msal
microsoft-agents-activity

# Note: Azure Developer CLI (azd) is not available via pip
# Run setup_deployment.sh or install manually with:
# curl -fsSL https://aka.ms/install-azd.sh | bash